requirements.txt
fastapi==0.88.0
pydantic==1.10.2
gunicorn
uvicorn
watchfiles
python-dotenv
motor